The server accepts games
The uploaded game can be accessed through the browser, but
the way it is served will probably be changed. Right now
nginx serves the whole static/ folder, and the game can be
accessed by manually typing:
   localhost/static/gamedata/latkazombit/kiigame.html

Also, better tests and comments should be added before accepting
this feature into the main branches.