Uploaded games are listed, are accessible, and work!
The uploaded games are listed on the front page with the help of
the (Flask built-in) jinja2 templating engine. The links point to
<name_of_game>/<some>.html, and the text is just <name_of_game>.
The name of the game is extracted from the folder structure,
specifically it is the name of the folder the .html file was found
from.

A follow-up task would be removing .html files from the list of
allowed extensions, and instead use a template to create the
game.html file.