Serving common files (javascript, images, ..)
Javascript files are automatically served from
static/commons/javascript. If some other kind of file is not
found from the game folder, the server will try to serve it from
commons/art folder. This is obviously mainly for images and sound
files, which may or may not be shared between different games.

One possible feature is that whenever images or sound effects
are uploaded to the server, they are also put into the commons/art
folder. When the user is using the editor he should add any assets
to the games (local) folder and THEN link to them with the editor -
if he does this the assets will be uploaded to the server.
However since this is not enforced currently, he may link to assets
that are located elsewhere, in which case they won't be uploaded.