Fixed flaw from generating the game.html
Previously the game.html was generated when the client
was attempting to upload a html file. This obviously doesn't
work, since html files should not be uploaded by the editor.
(The editor only uploads json, images, and music).
The editor also send the path of the game as a variable, from that
it's easier to extract the name.